requirements.txt
flask
psycopg2-binary
requests
redis